+++ geolocator_android/android/src/main/java/com/baseflow/geolocator/MethodCallHandlerImpl.java
@@ -126,7 +126,8 @@ class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
 
   private void onCheckPermission(MethodChannel.Result result) {
     try {
-      LocationPermission permission = this.permissionManager.checkPermissionStatus(context);
+      LocationPermission permission =
+          this.permissionManager.checkPermissionStatus(context);
       result.success(permission.toInt());
     } catch (PermissionUndefinedException e) {
       ErrorCodes errorCode = ErrorCodes.permissionDefinitionsNotFound;
@@ -164,11 +165,20 @@ class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
   }
 
   private void onGetLastKnownPosition(MethodCall call, MethodChannel.Result result) {
-    Boolean forceLocationManager = call.argument("forceAndroidLocationManager");
+    try{
+        if (!permissionManager.hasPermission(this.context)){
+            result.error(ErrorCodes.permissionDenied.toString(),ErrorCodes.permissionDenied.toDescription(),null);
+            return;
+        }
+    } catch (PermissionUndefinedException e) {
+        result.error(ErrorCodes.permissionDefinitionsNotFound.toString(), ErrorCodes.permissionDefinitionsNotFound.toDescription(), null);
+        return;
+    }
+
+    Boolean forceLocationManager = call.argument("forceLocationManager");
 
     this.geolocationManager.getLastKnownPosition(
         this.context,
-        this.activity,
         forceLocationManager != null && forceLocationManager,
         (Location location) -> result.success(LocationMapper.toHashMap(location)),
         (ErrorCodes errorCode) ->
@@ -176,7 +186,17 @@ class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
   }
 
   private void onGetCurrentPosition(MethodCall call, MethodChannel.Result result) {
-    Boolean forceLocationManager = call.argument("forceAndroidLocationManager");
+    try{
+          if (!permissionManager.hasPermission(this.context)){
+           result.error(ErrorCodes.permissionDenied.toString(),ErrorCodes.permissionDenied.toDescription(),null);
+           return;
+         }
+    } catch (PermissionUndefinedException e) {
+       result.error(ErrorCodes.permissionDefinitionsNotFound.toString(), ErrorCodes.permissionDefinitionsNotFound.toDescription(), null);
+       return;
+    }
+
+    Boolean forceLocationManager = call.argument("forceLocationManager");
     @SuppressWarnings("unchecked")
     Map<String, Object> map = (Map<String, Object>) call.arguments;
     LocationOptions locationOptions = LocationOptions.parseArguments(map);
@@ -187,9 +207,8 @@ class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
             this.context, forceLocationManager != null && forceLocationManager, locationOptions);
 
     geolocationManager.startPositionUpdates(
-        this.context,
-        this.activity,
         locationClient,
+        this.activity,
         (Location location) -> {
           if (replySubmitted[0]) {
             return;