+++ geolocator_android/android/src/main/java/com/baseflow/geolocator/StreamHandlerImpl.java
@@ -6,10 +6,13 @@ import android.location.Location;
 import android.util.Log;
 import androidx.annotation.Nullable;
 import com.baseflow.geolocator.errors.ErrorCodes;
+import com.baseflow.geolocator.errors.PermissionUndefinedException;
 import com.baseflow.geolocator.location.GeolocationManager;
 import com.baseflow.geolocator.location.LocationClient;
 import com.baseflow.geolocator.location.LocationMapper;
 import com.baseflow.geolocator.location.LocationOptions;
+import com.baseflow.geolocator.permission.PermissionManager;
+
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.EventChannel;
 
@@ -19,14 +22,16 @@ class StreamHandlerImpl implements EventChannel.StreamHandler {
   private static final String TAG = "StreamHandlerImpl";
 
   private final GeolocationManager geolocationManager;
+  private final PermissionManager permissionManager;
 
   @Nullable private EventChannel channel;
   @Nullable private Context context;
   @Nullable private Activity activity;
   @Nullable private LocationClient locationClient;
 
-  public StreamHandlerImpl(GeolocationManager geolocationManager) {
+  public StreamHandlerImpl(GeolocationManager geolocationManager, PermissionManager permissionManager) {
     this.geolocationManager = geolocationManager;
+    this.permissionManager = permissionManager;
   }
 
   void setActivity(@Nullable Activity activity) {
@@ -68,10 +73,20 @@ class StreamHandlerImpl implements EventChannel.StreamHandler {
 
   @Override
   public void onListen(Object arguments, EventChannel.EventSink events) {
+    try {
+        if (!permissionManager.hasPermission(this.context)){
+            events.error(ErrorCodes.permissionDenied.toString(),ErrorCodes.permissionDenied.toDescription(),null);
+            return;
+        }
+    } catch (PermissionUndefinedException e) {
+        events.error(ErrorCodes.permissionDefinitionsNotFound.toString(), ErrorCodes.permissionDefinitionsNotFound.toDescription(), null);
+        return; 
+    }
+    
     @SuppressWarnings("unchecked")
     Map<String, Object> map = (Map<String, Object>) arguments;
 
-    boolean forceLocationManager = (boolean) map.get("forceAndroidLocationManager");
+    boolean forceLocationManager = (boolean) map.get("forceLocationManager");
     LocationOptions locationOptions = LocationOptions.parseArguments(map);
 
     this.locationClient =
@@ -79,9 +94,8 @@ class StreamHandlerImpl implements EventChannel.StreamHandler {
             this.context, forceLocationManager, locationOptions);
 
     geolocationManager.startPositionUpdates(
-        context,
-        activity,
         this.locationClient,
+        activity,
         (Location location) -> events.success(LocationMapper.toHashMap(location)),
         (ErrorCodes errorCodes) ->
             events.error(errorCodes.toString(), errorCodes.toDescription(), null));