+++ geolocator_android/lib/geolocator_android.dart
@@ -1,0 +1,1 @@
+export 'src/types/android_settings.dart';