+++ geolocator_apple/apple/Classes/Handlers/PositionStreamHandler.m
@@ -5,31 +5,29 @@
 //  Created by Maurits van Beusekom on 04/06/2021.
 //
 
+#import "PermissionHandler.h"
 #import "PositionStreamHandler.h"
 #import "../Constants/ErrorCodes.h"
+#import "../Utils/ActivityTypeMapper.h"
 #import "../Utils/LocationAccuracyMapper.h"
 #import "../Utils/LocationDistanceMapper.h"
 #import "../Utils/LocationMapper.h"
-#import "../Utils/PermissionUtils.h"
 
 @interface PositionStreamHandler()
 @property (strong, nonatomic, nonnull) GeolocationHandler *geolocationHandler;
-@property (strong, nonatomic, nonnull) PermissionHandler *permissionHandler;
 @end
 
 @implementation PositionStreamHandler {
   FlutterEventSink _eventSink;
 }
 
-- (id) initWithGeolocationHandler: (GeolocationHandler * _Nonnull)geolocationHandler
-                PermissionHandler: (PermissionHandler * _Nonnull)permissionHandler {
+- (id) initWithGeolocationHandler: (GeolocationHandler * _Nonnull)geolocationHandler {
   self = [super init];
   if (!self) {
     return nil;
   }
   
   self.geolocationHandler = geolocationHandler;
-  self.permissionHandler = permissionHandler;
   
   return self;
 }
@@ -42,45 +40,35 @@
                                message: @"Already listening for location updates. If you want to restart listening please cancel other subscriptions first."
                                details: nil];
   }
+
   _eventSink = eventSink;
-  
+    
   __weak typeof(self) weakSelf = self;
   
-  [_permissionHandler
-   requestPermission:^(CLAuthorizationStatus status) {
-    
-    if (![PermissionUtils isStatusGranted:status]) {
-      [weakSelf onLocationFailureWithErrorCode: GeolocatorErrorPermissionDenied
-                              errorDescription: @"User denied permissions to access the device's location."];
-      return;
-    }
-    
-    CLLocationAccuracy accuracy = [LocationAccuracyMapper toCLLocationAccuracy:(NSNumber *)arguments[@"accuracy"]];
-    CLLocationDistance distanceFilter = [LocationDistanceMapper toCLLocationDistance:(NSNumber *)arguments[@"distanceFilter"]];
-    
-    [[weakSelf geolocationHandler] startListeningWithDesiredAccuracy:accuracy
-                                                      distanceFilter:distanceFilter
-                                                       resultHandler:^(CLLocation *location) {
-      [weakSelf onLocationDidChange: location];
-    }
-                                                        errorHandler:^(NSString *errorCode, NSString *errorDescription){
-      [weakSelf onLocationFailureWithErrorCode:errorCode
-                              errorDescription:errorDescription];
-    }];
+  CLLocationAccuracy accuracy = [LocationAccuracyMapper toCLLocationAccuracy:(NSNumber *)arguments[@"accuracy"]];
+  CLLocationDistance distanceFilter = [LocationDistanceMapper toCLLocationDistance:(NSNumber *)arguments[@"distanceFilter"]];
+  NSNumber* pauseLocationUpdatesAutomatically = arguments[@"pauseLocationUpdatesAutomatically"];
+  CLActivityType activityType = [ActivityTypeMapper toCLActivityType:(NSNumber *)arguments[@"activityType"]];
+	      
+  [[weakSelf geolocationHandler] startListeningWithDesiredAccuracy:accuracy
+                                                    distanceFilter:distanceFilter
+                                 pauseLocationUpdatesAutomatically:pauseLocationUpdatesAutomatically && [pauseLocationUpdatesAutomatically boolValue]
+                                                      activityType:activityType
+                                                     resultHandler:^(CLLocation *location) {
+    [weakSelf onLocationDidChange: location];
   }
-   errorHandler:^(NSString *errorCode, NSString *errorDescription) {
+                                                      errorHandler:^(NSString *errorCode, NSString *errorDescription){
     [weakSelf onLocationFailureWithErrorCode:errorCode
                             errorDescription:errorDescription];
   }];
-  
   return nil;
 }
 
-- (FlutterError *)onCancelWithArguments:(id)arguments{
-  [_geolocationHandler stopListening];
-  _eventSink = nil;
-  
-  return nil;
+- (FlutterError *_Nullable)onCancelWithArguments:(id _Nullable)arguments {
+    [_geolocationHandler stopListening];
+    _eventSink = nil;
+    
+    return nil;
 }
 
 - (void)onLocationDidChange:(CLLocation *_Nullable)location {
@@ -96,7 +84,6 @@
   _eventSink([FlutterError errorWithCode:errorCode
                                  message:errorDescription
                                  details:nil]);
-  _eventSink = nil;
 }
 
 @end