+++ geolocator_apple/lib/geolocator_apple.dart
@@ -1,0 +1,2 @@
+export 'src/types/activity_type.dart';
+export 'src/types/apple_settings.dart';