+++ geolocator_platform_interface/lib/src/implementations/method_channel_geolocator.dart
@@ -7,7 +7,6 @@ import '../enums/enums.dart';
 import '../errors/errors.dart';
 import '../extensions/extensions.dart';
 import '../geolocator_platform_interface.dart';
-import '../models/location_options.dart';
 import '../models/position.dart';
 
 /// An implementation of [GeolocatorPlatform] that uses method channels.
@@ -44,9 +43,9 @@ class MethodChannelGeolocator extends GeolocatorPlatform {
 
       return permission.toLocationPermission();
     } on PlatformException catch (e) {
-      _handlePlatformException(e);
+      final error = _handlePlatformException(e);
 
-      rethrow;
+      throw error;
     }
   }
 
@@ -59,9 +58,9 @@ class MethodChannelGeolocator extends GeolocatorPlatform {
 
       return permission.toLocationPermission();
     } on PlatformException catch (e) {
-      _handlePlatformException(e);
+      final error = _handlePlatformException(e);
 
-      rethrow;
+      throw error;
     }
   }
 
@@ -84,9 +83,9 @@ class MethodChannelGeolocator extends GeolocatorPlatform {
 
       return positionMap != null ? Position.fromMap(positionMap) : null;
     } on PlatformException catch (e) {
-      _handlePlatformException(e);
+      final error = _handlePlatformException(e);
 
-      rethrow;
+      throw error;
     }
   }
 
@@ -99,38 +98,33 @@ class MethodChannelGeolocator extends GeolocatorPlatform {
 
   @override
   Future<Position> getCurrentPosition({
-    LocationAccuracy desiredAccuracy = LocationAccuracy.best,
-    bool forceAndroidLocationManager = false,
-    Duration? timeLimit,
+    LocationSettings? locationSettings,
   }) async {
-    final locationOptions = LocationOptions(
-      accuracy: desiredAccuracy,
-      forceAndroidLocationManager: forceAndroidLocationManager,
-    );
-
     try {
       Future<dynamic> positionFuture;
+
+      var timeLimit = locationSettings?.timeLimit;
 
       if (timeLimit != null) {
         positionFuture = _methodChannel
             .invokeMethod(
               'getCurrentPosition',
-              locationOptions.toJson(),
+              locationSettings?.toJson(),
             )
             .timeout(timeLimit);
       } else {
         positionFuture = _methodChannel.invokeMethod(
           'getCurrentPosition',
-          locationOptions.toJson(),
+          locationSettings?.toJson(),
         );
       }
 
       final positionMap = await positionFuture;
       return Position.fromMap(positionMap);
     } on PlatformException catch (e) {
-      _handlePlatformException(e);
+      final error = _handlePlatformException(e);
 
-      rethrow;
+      throw error;
     }
   }
 
@@ -147,7 +141,7 @@ class MethodChannelGeolocator extends GeolocatorPlatform {
         .handleError((error) {
       _serviceStatusStream = null;
       if (error is PlatformException) {
-        _handlePlatformException(error);
+        error = _handlePlatformException(error);
       }
       throw error;
     });
@@ -157,25 +151,17 @@ class MethodChannelGeolocator extends GeolocatorPlatform {
 
   @override
   Stream<Position> getPositionStream({
-    LocationAccuracy desiredAccuracy = LocationAccuracy.best,
-    int distanceFilter = 0,
-    bool forceAndroidLocationManager = false,
-    int timeInterval = 0,
-    Duration? timeLimit,
+    LocationSettings? locationSettings,
   }) {
-    final locationOptions = LocationOptions(
-      accuracy: desiredAccuracy,
-      distanceFilter: distanceFilter,
-      forceAndroidLocationManager: forceAndroidLocationManager,
-      timeInterval: timeInterval,
-    );
     if (_positionStream != null) {
       return _positionStream!;
     }
     var originalStream = _eventChannel.receiveBroadcastStream(
-      locationOptions.toJson(),
+      locationSettings?.toJson(),
     );
     var positionStream = _wrapStream(originalStream);
+
+    var timeLimit = locationSettings?.timeLimit;
 
     if (timeLimit != null) {
       positionStream = positionStream.timeout(
@@ -196,9 +182,8 @@ class MethodChannelGeolocator extends GeolocatorPlatform {
             Position.fromMap(element.cast<String, dynamic>()))
         .handleError(
       (error) {
-        _positionStream = null;
         if (error is PlatformException) {
-          _handlePlatformException(error);
+          error = _handlePlatformException(error);
         }
         throw error;
       },
@@ -226,8 +211,8 @@ class MethodChannelGeolocator extends GeolocatorPlatform {
       );
       return LocationAccuracyStatus.values[status];
     } on PlatformException catch (e) {
-      _handlePlatformException(e);
-      rethrow;
+      final error = _handlePlatformException(e);
+      throw error;
     }
   }
 
@@ -241,24 +226,24 @@ class MethodChannelGeolocator extends GeolocatorPlatform {
       .invokeMethod<bool>('openLocationSettings')
       .then((value) => value ?? false);
 
-  void _handlePlatformException(PlatformException exception) {
+  Exception _handlePlatformException(PlatformException exception) {
     switch (exception.code) {
       case 'ACTIVITY_MISSING':
-        throw ActivityMissingException(exception.message);
+        return ActivityMissingException(exception.message);
       case 'LOCATION_SERVICES_DISABLED':
-        throw const LocationServiceDisabledException();
+        return const LocationServiceDisabledException();
       case 'LOCATION_SUBSCRIPTION_ACTIVE':
-        throw const AlreadySubscribedException();
+        return const AlreadySubscribedException();
       case 'PERMISSION_DEFINITIONS_NOT_FOUND':
-        throw PermissionDefinitionsNotFoundException(exception.message);
+        return PermissionDefinitionsNotFoundException(exception.message);
       case 'PERMISSION_DENIED':
-        throw PermissionDeniedException(exception.message);
+        return PermissionDeniedException(exception.message);
       case 'PERMISSION_REQUEST_IN_PROGRESS':
-        throw PermissionRequestInProgressException(exception.message);
+        return PermissionRequestInProgressException(exception.message);
       case 'LOCATION_UPDATE_FAILURE':
-        throw PositionUpdateException(exception.message);
+        return PositionUpdateException(exception.message);
       default:
-        throw exception;
+        return exception;
     }
   }
 }