+++ geolocator_apple/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 2.0.0+2
+
+- Removes the Android specific `timeInterval` parameter from `AppleSettings`.
+
+## 2.0.0+1
+
+- Adds missing `timeLimit` to the `AppleSettings` class.
+
+## 2.0.0
+
+- iOS: Keep `PositionStream` alive when the `Location Services` has been turned off and on again in the settings.
+- Removed implicit request for permissions when getting a position.
+- Added the [ActivityType] enum needed for the `pauseLocationUpdatesAutomatically` property.
+- Added the `pauseLocationUpdatesAutomatically` and `activityType` property to the iOS options class.
+
 ## 1.2.2
 
 - Fixed iOS cancelation of positionStream.